feat: InputBoardTitle 컴포넌트 분리
